New changes imported
There have been changes made on minor functionality details(play picture changes on click, add button doesn't generate notes when triggered if no element selected to add), as well as New, Open, Save and Save As options added.  MessageBox added for Instructions in the Help menuItem but does not work properly (it blocks the form but it doesn't show anywhere. Haven't tried to see if it is in the background of the form.)